Restructured folders and added animation for shooting shotgun
