+++ Levels/Scenes/test_level.tscn
@@ -3,13 +3,13 @@
 [ext_resource type="PackedScene" uid="uid://bnpdedb212get" path="res://Levels/Scenes/level_end.tscn" id="1_2e8qi"]
 [ext_resource type="Texture2D" uid="uid://cu23mts0uo7fc" path="res://Assets/Tilemaps/garden_tilemap_v3.png" id="1_ohl1y"]
 [ext_resource type="PackedScene" uid="uid://dv1ycuaoyk5yv" path="res://Items/NodeScenes/items.tscn" id="3_4skx5"]
-[ext_resource type="PackedScene" uid="uid://dlyfs7e3himjx" path="res://Player/PlayerCharacter/Scenes/player_cat.tscn" id="3_tgiyf"]
+[ext_resource type="PackedScene" uid="uid://dlyfs7e3himjx" path="res://Players/PlayerCharacter/Scenes/player_cat.tscn" id="3_tgiyf"]
 [ext_resource type="PackedScene" uid="uid://77tyaj4jl48x" path="res://Items/NodeScenes/items_coffee.tscn" id="4_5r06v"]
 [ext_resource type="PackedScene" uid="uid://dpk3inegf0yit" path="res://Items/Scenes/coffee.tscn" id="5_igh1l"]
 [ext_resource type="PackedScene" uid="uid://wf6m2blo1d8u" path="res://Items/NodeScenes/items_coffee_pot.tscn" id="6_w2ons"]
 [ext_resource type="PackedScene" uid="uid://f41uy40r0hsp" path="res://Items/Scenes/coffee_pot.tscn" id="7_lp4a7"]
-[ext_resource type="PackedScene" uid="uid://xgbwu1var27k" path="res://Player/Weapons/Pickups/Scenes/weapon_pickup_shotgun.tscn" id="10_dyv2w"]
-[ext_resource type="PackedScene" uid="uid://0e8r4l08v0gy" path="res://Player/Weapons/Pickups/Scenes/weapon_pickup_revolver.tscn" id="10_m8g7i"]
+[ext_resource type="PackedScene" uid="uid://xgbwu1var27k" path="res://Players/Weapons/Pickups/Scenes/weapon_pickup_shotgun.tscn" id="10_dyv2w"]
+[ext_resource type="PackedScene" uid="uid://0e8r4l08v0gy" path="res://Players/Weapons/Pickups/Scenes/weapon_pickup_revolver.tscn" id="10_m8g7i"]
 
 [sub_resource type="TileSetAtlasSource" id="TileSetAtlasSource_iy0cb"]
 texture = ExtResource("1_ohl1y")
