--- Player/Components/Scenes/shooting_component.tscn
@@ -1,6 +1,0 @@
-[gd_scene load_steps=2 format=3 uid="uid://ciag6jxw7f4v8"]
-
-[ext_resource type="Script" path="res://Player/Components/Scripts/ShootingComponent.cs" id="1_60ab8"]
-
-[node name="shooting_component" type="Node"]
-script = ExtResource("1_60ab8")
