--- Player/WeaponManager/Scenes/weapon_manager.tscn
@@ -1,6 +1,0 @@
-[gd_scene load_steps=2 format=3 uid="uid://dmbsjapq5hdtm"]
-
-[ext_resource type="Script" path="res://Player/WeaponManager/Scripts/WeaponManagerScript.cs" id="1_li4jb"]
-
-[node name="weapon_manager" type="Node"]
-script = ExtResource("1_li4jb")
