--- Player/Weapons/Shotgun/Scenes/bullet_shotgun.tscn
@@ -1,97 +1,0 @@
-[gd_scene load_steps=14 format=3 uid="uid://b8lo3du126a5h"]
-
-[ext_resource type="Script" path="res://Player/Weapons/Shotgun/Scripts/BulletShotgun.cs" id="1_xs1l8"]
-[ext_resource type="Texture2D" uid="uid://8qngepld1gb2" path="res://Assets/Sprites/PlayerSprites/cat_shotgun-3.png" id="2_nijsl"]
-[ext_resource type="Texture2D" uid="uid://rwk8vjrh3xrp" path="res://Assets/Sprites/PlayerSprites/cat_shotgun-2.png" id="3_st25n"]
-
-[sub_resource type="AtlasTexture" id="AtlasTexture_v7lit"]
-atlas = ExtResource("2_nijsl")
-region = Rect2(0, 512, 48, 32)
-
-[sub_resource type="AtlasTexture" id="AtlasTexture_oxyoa"]
-atlas = ExtResource("2_nijsl")
-region = Rect2(48, 512, 48, 32)
-
-[sub_resource type="AtlasTexture" id="AtlasTexture_lcma7"]
-atlas = ExtResource("3_st25n")
-region = Rect2(0, 256, 48, 32)
-
-[sub_resource type="AtlasTexture" id="AtlasTexture_0eldk"]
-atlas = ExtResource("3_st25n")
-region = Rect2(48, 256, 48, 32)
-
-[sub_resource type="AtlasTexture" id="AtlasTexture_whwa3"]
-atlas = ExtResource("3_st25n")
-region = Rect2(96, 256, 48, 32)
-
-[sub_resource type="AtlasTexture" id="AtlasTexture_uide2"]
-atlas = ExtResource("3_st25n")
-region = Rect2(144, 256, 48, 32)
-
-[sub_resource type="AtlasTexture" id="AtlasTexture_ggepx"]
-atlas = ExtResource("3_st25n")
-region = Rect2(192, 256, 48, 32)
-
-[sub_resource type="SpriteFrames" id="SpriteFrames_vb4a3"]
-animations = [{
-"frames": [{
-"duration": 1.0,
-"texture": SubResource("AtlasTexture_v7lit")
-}, {
-"duration": 1.0,
-"texture": SubResource("AtlasTexture_oxyoa")
-}],
-"loop": true,
-"name": &"fly",
-"speed": 15.0
-}, {
-"frames": [{
-"duration": 1.0,
-"texture": SubResource("AtlasTexture_lcma7")
-}, {
-"duration": 1.0,
-"texture": SubResource("AtlasTexture_0eldk")
-}, {
-"duration": 1.0,
-"texture": SubResource("AtlasTexture_whwa3")
-}, {
-"duration": 1.0,
-"texture": SubResource("AtlasTexture_uide2")
-}, {
-"duration": 1.0,
-"texture": SubResource("AtlasTexture_ggepx")
-}],
-"loop": false,
-"name": &"hit",
-"speed": 18.0
-}]
-
-[sub_resource type="RectangleShape2D" id="RectangleShape2D_fb027"]
-size = Vector2(14.1944, 10.1731)
-
-[sub_resource type="RectangleShape2D" id="RectangleShape2D_mymv6"]
-size = Vector2(3.77778, 10.7692)
-
-[node name="bullet_shotgun" type="Area2D"]
-collision_layer = 128
-collision_mask = 81
-script = ExtResource("1_xs1l8")
-
-[node name="sprite" type="AnimatedSprite2D" parent="."]
-scale = Vector2(0.35, 0.35)
-sprite_frames = SubResource("SpriteFrames_vb4a3")
-animation = &"fly"
-
-[node name="collider" type="CollisionShape2D" parent="."]
-scale = Vector2(0.45, 0.45)
-shape = SubResource("RectangleShape2D_fb027")
-
-[node name="hitbox" type="Area2D" parent="."]
-collision_layer = 2
-collision_mask = 81
-
-[node name="hitbox_collider" type="CollisionShape2D" parent="hitbox"]
-position = Vector2(-3.85, 0)
-scale = Vector2(0.45, 0.45)
-shape = SubResource("RectangleShape2D_mymv6")
-debug_color = Color(0.956307, 0.130976, 0.417223, 0.42)
