+++ Players/WeaponManager/Scenes/weapon_manager.tscn
@@ -1,0 +1,6 @@
+[gd_scene load_steps=2 format=3 uid="uid://dmbsjapq5hdtm"]
+
+[ext_resource type="Script" path="res://Players/WeaponManager/Scripts/WeaponManagerScript.cs" id="1_li4jb"]
+
+[node name="weapon_manager" type="Node"]
+script = ExtResource("1_li4jb")
